replace all importas -i with importas -D
importas -i is way too strict and results in hard failures if the
variable isn't imported in the environment. This can be bad in some
instances for users (e.g. a .conf file changes and a user forgets to
incorporate the changes) and in general it is better to just have safe
fallbacks. Additionally, there were some other minor erros fixed in here
such as incorrect variable names and such (some have been wrong for a
long time but no one complained which shows you how much people use some
of these services).