wpa_supplicant: add a dhcpcd dependency
wpa_supplicant needs dhcpcd running in order to actually connect to any
wireless interface. This used to have a dependency in the past, but this
was removed because of some weird behavior with dhcpcd not properly
shutting down. This was taken care of with the last commit though so we
can add this dependency back.